Netatmo welcome camera initial implementation
